Update EDAMAME Posture version to 0.9.82 and modify Chocolatey installation script in GitHub Actions workflow
This commit updates the version of the EDAMAME Posture package to 0.9.82 in both the nuspec file and the Chocolatey installation script. It also adjusts the installation script to dynamically set the package version and URL, ensuring that the correct version is used during installation. These changes enhance the package management process and improve the reliability of the deployment workflow.